Implemented basic paging for the game
Implemented a Page abstract class. This class represents one unit
of displaying content to the user and receiving input. Furthermore,
pages are assumed to be contained in a stack of pages, where the
top page is currently interacting with the user. A page can modify
this stack by returning a StackCommand after receiving player input.
The current stack commands are null (so keep this page active), pop,
and replace.